Use f64 end-to-end to match Sinumerik REAL precision (#23)
* Use f64 end-to-end to match Sinumerik REAL precision

Sinumerik program data of type REAL is a 64-bit IEEE double (NC
programming manual 4.1.3.6); R parameters and system variables are
TYPE_DOUBLE, and comparison commands are specified with a relative
equality of 1e-12 - none of which is representable in f32 (~7
significant digits, relative epsilon ~1.2e-7).

Migrate Value, State (axes/symbol_table/translation) and the whole
expression evaluator from f32 to f64. DIV now divides the REAL
operands and truncates the result (manual 4.1.3.1: 7 DIV 4.1 = 1)
instead of truncating the operands through integer casts, which also
removes the bogus division-by-zero error for divisors below 1.
Output DataFrame columns become Float64 natively instead of being
cast from f32-quantized values, so emitted coordinates are now
bit-exact for decimal literals.

Closes #22


* Pin DIV result-truncation semantics in goldens; fix stale README dtypes

7 DIV 4.1 = 1 and 7 DIV 0.4 = 17 per manual 4.1.3.1, and the README
example output now shows the actual Float64 columns.


* Polish from Copilot review comments across the stack

- frame_assignments: snapshot the axis state once per frame
  instruction instead of cloning the map for every assignment
- FRAME_KEYWORDS check uses eq_ignore_ascii_case instead of
  allocating an uppercased String per G-command
- Correct the normalize_reserved_case doc example: the bare word
  form X10 is uppercase-only in the grammar; case-insensitivity
  applies to the x=10 form
- Remove stray blank lines before the DataFrame construction

The remaining Copilot comments on #18 (axis/block-address case
normalization, Display-formatted Python errors) were already fixed
in #20.


* Address Copilot review on #23: comparisons, literal parsing, N lexeme

- REAL comparisons now use relative equality with a 1e-12 tolerance per
  the NC programming manual ("Precision correction on comparison errors
  (TRUNC)"): ==, <>, <= and >= treat relatively-equal values as equal,
  and > / < exclude them (the MD10280 Bit0=0 default). Pinned with a
  parametrized pytest.
- Numeric literal parsing returns a line-annotated ParsingError instead
  of panicking the whole interpreter via expect().
- Block numbers keep their original integer lexeme instead of round-
  tripping through f64, so large N values survive exactly; the output
  column parses as i64 first with a float fallback.


---------

+++ src/lib.rs
@@ -41,7 +41,7 @@ mod python_bindings {
     ) -> PyResult<(
         Py<PyDict>,
         Vec<(String, String)>,
-        HashMap<String, HashMap<String, f32>>,
+        HashMap<String, HashMap<String, f64>>,
     )> {
         let (table, state) = nc_to_table(
             input,
